Add in Badrobot a Kubernetes Operator audit tool (#34)
Badrobot is a Kubernetes Operator audit tool for Kubernetes Operators.
It statically analyses manifests for high risk configurations such as lack of security restrictions on the deployed controller and the permissions of an associated clusterole. The risk analysis is primarily focussed on the likelihood that a compromised Operator would be able to obtain full cluster permissions.

Currently 2 Releases latest v0.1.2, 5 Contributors, 25 Stars, Last commit 2 days ago.
Apache License 2.0 - https://github.com/controlplaneio/badrobot/blob/master/LICENSE

It is still in the early stages of this development but its every useful as demoed at KubeCon + CloudNativeCon Europe 2022.